+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- Don't panic on invalid authorization config. Previously, a missing OPA ConfigMap would crash the operator ([#XXX]).
 - getting_started: Add a 120 second timeout before trying to enable the DAG ([#665]).
 
 [#665]: https://github.com/stackabletech/airflow-operator/pull/665